decouple computation of dimensions
